+++ trans_qt_encode01/trans_qt_encode/encode/probility_estimation.c
@@ -37,12 +37,14 @@
 #define uint unsigned int
 #define uchar unsigned char
 
-void probility_estimation(Uint32_Dat* rk, Uint32_Dat* nk, float *prob)
+void probility_estimation(Uint32_Dat* rk, Uint32_Dat* nk, double *prob)
 {
 	int sum_rk = 0;
 	int max_rk = 0;
-	float prb1 = 0;
-	float prb2 = 0;
+	//float prb1 = 0;
+	//float prb2 = 0;
+	double prb1 = 0;
+	double prb2 = 0;
 	int thd = 0;
 
 
@@ -53,7 +55,7 @@ void probility_estimation(Uint32_Dat* rk, Uint32_Dat* nk, float *prob)
 		//printf("%d:%d sum_rk = %d \n", i+1, rk->dat[i], sum_rk);
 	}
 	/* prb2=length(rk)/sum(rk); */
-	prb2 = (float)rk->len / sum_rk;
+	prb2 = (double)rk->len / ((double)sum_rk);
 	//printf("prb2 = %f\n", prb2);
 	/* thd=max(rk); */
 	thd = max_rk;
@@ -77,3 +79,43 @@ void probility_estimation(Uint32_Dat* rk, Uint32_Dat* nk, float *prob)
 
 	return;
 }
+//void probility_estimation(Uint32_Dat* rk, Uint32_Dat* nk, float *prob)
+//{
+//	int sum_rk = 0;
+//	int max_rk = 0;
+//	float prb1 = 0;
+//	float prb2 = 0;
+//	int thd = 0;
+//
+//
+//	sum_rk = 0;
+//	for (int i = 0; i < rk->len; i++) {
+//		if (max_rk < rk->dat[i])  max_rk = rk->dat[i];
+//		sum_rk += rk->dat[i];
+//		//printf("%d:%d sum_rk = %d \n", i+1, rk->dat[i], sum_rk);
+//	}
+//	/* prb2=length(rk)/sum(rk); */
+//	prb2 = (float)rk->len / sum_rk;
+//	//printf("prb2 = %f\n", prb2);
+//	/* thd=max(rk); */
+//	thd = max_rk;
+//	/* if thd <=7 && length(rk)>360 */
+//	if (thd <= 7 && rk->len > 360) {
+//		/* prb1=prob_est(nk); */
+//		prob_est(nk, &prb1);
+//		/* prob=prb1; */
+//		*prob = prb1;
+//	}
+//	else if (thd <= 12) {  /* elseif thd<=12 */
+//	   /* prb1=prob_est(nk);  */
+//		prob_est(nk, &prb1);
+//		/* prob=(prb1+prb2)/2; */
+//		*prob = (prb1 + prb2) / 2;
+//	}
+//	else {
+//		/* prob=prb2; */
+//		*prob = prb2;
+//	}
+//
+//	return;
+//}
